feat: Trying to solve problem with spring starting
