+++ kotiki-java/data-access-layer/src/main/java/DataModule.java
@@ -1,11 +1,3 @@
-import breeds.Breed;
-import cats.Cat;
-import owner.Owner;
-import org.hibernate.SessionFactory;
-import org.hibernate.cfg.Configuration;
-
-import java.util.GregorianCalendar;
-
 public class DataModule {
     public static void main(String[] args) {
     }
