--- kotiki-java/data-access-layer/src/main/java/cats/ICatsDao.java
@@ -1,7 +1,0 @@
-package cats;
-
-import common.IDao;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ICatsDao extends IDao<Cat> {
-}
