+++ kotiki-java/data-access-layer/src/main/java/dao/ICatsDao.java
@@ -1,0 +1,10 @@
+package dao;
+
+import entities.Cat;
+import dao.IDao;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ICatsDao extends IDao<Cat> {
+}
