+++ kotiki-java/data-access-layer/src/main/java/entities/BaseEntity.java
@@ -1,7 +1,6 @@
-package common;
+package entities;
 
-import javax.persistence.Inheritance;
-import javax.persistence.InheritanceType;
+import javax.persistence.*;
 
 @Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
 public abstract class BaseEntity {
