+++ kotiki-java/data-access-layer/src/main/java/entities/Breed.java
@@ -1,4 +1,4 @@
-package breeds;
+package entities;
 
 public enum Breed {
     Abyssian,
