+++ kotiki-java/service-layer/src/main/java/com/kotiki/Main.java
@@ -1,4 +1,4 @@
-package application;
+package com.kotiki;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
