+++ kotiki-java/service-layer/src/main/java/com/kotiki/config/Config.java
@@ -1,0 +1,10 @@
+package com.kotiki.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+@Configuration
+@Import(value={ConfigCats.class})
+public class Config {
+
+}
