+++ kotiki-java/service-layer/src/main/java/com/kotiki/config/ConfigCats.java
@@ -1,0 +1,12 @@
+package com.kotiki.config;
+
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@EnableJpaRepositories(basePackages={"dao"})
+@EntityScan(basePackages = {"entities"})
+public class ConfigCats {
+}
