+++ kotiki-java/service-layer/src/main/java/com/kotiki/dto/OwnerDTO.java
@@ -1,10 +1,9 @@
-package dto;
+package com.kotiki.dto;
 
-import cats.Cat;
+import entities.Cat;
 
 import java.util.ArrayList;
 import java.util.Calendar;
-import java.util.GregorianCalendar;
 import java.util.List;
 
 public class OwnerDTO {
