+++ kotiki-java/service-layer/src/main/java/com/kotiki/exceptions/IncorrectCatParamsException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package com.kotiki.exceptions;
 
 public class IncorrectCatParamsException extends Exception{
     public IncorrectCatParamsException(String message) {
