+++ kotiki-java/service-layer/src/main/java/com/kotiki/exceptions/IncorrectOwnerParamException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package com.kotiki.exceptions;
 
 public class IncorrectOwnerParamException extends Exception{
     public IncorrectOwnerParamException(String message) {
