+++ kotiki-java/service-layer/src/main/java/com/kotiki/exceptions/NotFoundByIdException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package com.kotiki.exceptions;
 
 public class NotFoundByIdException extends Exception{
     public NotFoundByIdException(String message) {
