+++ kotiki-java/service-layer/src/main/java/com/kotiki/utils/OwnerDtoMapping.java
@@ -1,7 +1,7 @@
-package utils;
+package com.kotiki.utils;
 
-import dto.OwnerDTO;
-import owner.Owner;
+import com.kotiki.dto.OwnerDTO;
+import entities.Owner;
 
 public class OwnerDtoMapping {
     public OwnerDTO mapToDTO(Owner owner) {
