+++ kotiki-java/service-layer/src/main/resources/application.properties
@@ -1,1 +1,4 @@
-spring.datasource.url
+spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
+spring.datasource.username=postgres
+spring.datasource.password=password
+spring.datasource.driver-class-name=org.postgresql.Driver
